% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{Get info data
Function which will allow to have information on a data set passed in parameter: dimensions and name of columns in a list}
\usage{
get_info_data(data)
}
\value{
list
}
\description{
Get info data
Function which will allow to have information on a data set passed in parameter: dimensions and name of columns in a list
}
\examples{

get_info_data(data = iris)
}
