% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{Get_mean_data}
\usage{
get_mean_data(data)
}
\value{
table
}
\description{
Will get a table with the average of all numeric variables in your dataset as parameter.
}
\examples{
get_mean_data(data = iris)
}
