% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_starwar_humans.R
\docType{data}
\name{starwar_humans}
\alias{starwar_humans}
\title{starwar_humans}
\format{
A data frame with  rows and  variables:
\describe{
}
}
\source{
Source
}
\usage{
starwar_humans
}
\description{
Description.
}
\keyword{datasets}
